only copy og dotfiles once move dotfiles to src dir improve setup scripts -> should be idempotent

+++ setup.sh
@@ -8,10 +8,11 @@
 
 ########## Variables
 
-# get script dir
+# get scrip
 DIR="$( cd "$( dirname "$0" )" && pwd )"
+SRC="$DIR"/src
 OLD="$DIR"/old
-LEN="$((${#DIR} + 2))"
+LEN="$((${#SRC} + 2))"
 CONFIG_DIR="$DIR"/config
 CONFIG_LEN="$((${#CONFIG_DIR} + 2))"
 ##########
@@ -23,36 +24,34 @@ rm -rf "$HOME"/dotfiles_old
 echo "Creating $OLD for backup of any existing dotfiles in ~"
 mkdir -p "$OLD"
 
-# move any existing dotfiles in homedir to dotfiles/old directory, then create symlinks
-for FILE in "$DIR"/*; do
-    # skip *.sh files and README.md
-    if [[ ! "$FILE" == *.sh ]] && [[ ! "$FILE" == *README.md ]] && [[ ! "$FILE" == "$OLD" ]] && [[ ! "$FILE" == *config ]]; then
-        BASE="$(echo "$FILE" | cut -c "$LEN"-)"
+# move any existing dotfiles in homedir to dotfiles/old directory, then create symlinks to dotfiles/src
+for FILE in "$SRC"/*; do
+  BASE="$(echo "$FILE" | cut -c "$LEN"-)"
 
-        if [ -f ~/."$BASE" ] || [ -d ~/."$BASE" ]; then
-          echo "Moving ~/.$BASE to $OLD"
-          cp -LR ~/."$BASE" "$OLD/$BASE"
-        fi
+  if ([ -f ~/."$BASE" ] || [ -d ~/."$BASE" ]) && ([ ! -f "$OLD/$BASE" ] && [ ! -d "$OLD/$BASE" ]) ; then
+    echo "Moving ~/.$BASE to $OLD"
+    cp -LR ~/."$BASE" "$OLD/$BASE"
+  fi
 
-        echo "Creating symlink to $FILE in home directory"
-        rm -rf ~/."$BASE" && ln -s "$FILE" ~/."$BASE"
-    fi
+  echo "Creating symlink to $FILE in home directory"
+  rm -rf ~/."$BASE" && ln -s "$SRC/$BASE" ~/."$BASE"
 done
 
+# move any existing files in ~/config to dotfiles/old/config then create symlinks
 mkdir -p ~/.config
 mkdir -p "$OLD"/config
 for FILE in "$DIR"/config/*; do
-    BASE="$(echo "$FILE" | cut -c "$CONFIG_LEN"-)"
-    echo ~/.config/"$BASE"
+  BASE="$(echo "$FILE" | cut -c "$CONFIG_LEN"-)"
 
-    if [ -f ~/.config/"$BASE" ] || [ -d ~/.config/"$BASE" ]; then
-        echo "Moving ~/.config/$BASE to $OLD"
-        cp -LR ~/.config/"$BASE" "$OLD"/config/
-    fi
+  if ([ -f ~/.config/"$BASE" ] || [ -d ~/.config/"$BASE" ]) && ([ ! -f "$OLD"/config/"$BASE" ] && [ ! -d "$OLD"/config/"$BASE" ]); then
+    echo "Moving ~/.config/$BASE to $OLD"
+    cp -LR ~/.config/"$BASE" "$OLD"/config/
+  fi
 
-    echo "Creating symlink to $FILE in ~/.config directory"
-    rm -rf ~/.config/"$BASE" && ln -s "$FILE" ~/.config/"$BASE"
+  echo "Creating symlink to $FILE in ~/.config directory"
+  rm -rf ~/.config/"$BASE" && ln -s "$FILE" ~/.config/"$BASE"
 done
 
-vim +PlugInstall +qall
-
+"$DIR"/config_vim.sh
+"$DIR"/install_fonts.sh
+"$DIR"/install_kitty.sh
